+++ umh-core/internal/fsmtest/benthos.go
@@ -184,110 +184,6 @@ func TransitionToBenthosState(mockService *benthossvc.MockBenthosService, servic
 	}
 }
 
-// WaitForBenthosManagerStable is a simple helper that calls manager.Reconcile once
-// and checks if there was an error. It's used for scenarios like an empty config
-// where we just want to do one pass. Then we might call it multiple times if needed.
-func WaitForBenthosManagerStable(
-	ctx context.Context,
-	manager *benthosfsm.BenthosManager,
-	fullCfg config.FullConfig,
-	startTick uint64,
-) (uint64, error) {
-	err, _ := manager.Reconcile(ctx, fullCfg, startTick)
-	if err != nil {
-		return startTick, err
-	}
-	// We might do more checks if needed, e.g. manager has zero instances?
-	return startTick + 1, nil
-}
-
-// WaitForBenthosManagerInstanceState repeatedly calls manager.Reconcile until
-// we see the specified 'desiredState' or we hit maxAttempts.
-func WaitForBenthosManagerInstanceState(
-	ctx context.Context,
-	manager *benthosfsm.BenthosManager,
-	fullCfg config.FullConfig,
-	instanceName string,
-	desiredState string,
-	maxAttempts int,
-	startTick uint64,
-) (uint64, error) {
-	tick := startTick
-	for i := 0; i < maxAttempts; i++ {
-		err, _ := manager.Reconcile(ctx, fullCfg, tick)
-		if err != nil {
-			return tick, err
-		}
-		tick++
-
-		inst, found := manager.GetInstance(instanceName)
-		if found && inst.GetCurrentFSMState() == desiredState {
-			return tick, nil
-		}
-	}
-	return tick, fmt.Errorf("instance %s did not reach state %s after %d attempts",
-		instanceName, desiredState, maxAttempts)
-}
-
-// WaitForBenthosManagerInstanceRemoval waits until the given serviceName is removed
-// from the manager's instance map.
-func WaitForBenthosManagerInstanceRemoval(
-	ctx context.Context,
-	manager *benthosfsm.BenthosManager,
-	fullCfg config.FullConfig,
-	instanceName string,
-	maxAttempts int,
-	startTick uint64,
-) (uint64, error) {
-	tick := startTick
-	for i := 0; i < maxAttempts; i++ {
-		err, _ := manager.Reconcile(ctx, fullCfg, tick)
-		if err != nil {
-			return tick, err
-		}
-		tick++
-
-		if _, found := manager.GetInstance(instanceName); !found {
-			// success
-			return tick, nil
-		}
-	}
-	return tick, fmt.Errorf("instance %s not removed after %d attempts", instanceName, maxAttempts)
-}
-
-// WaitForBenthosManagerMultiState can check multiple instances at once:
-// e.g. map[serviceName]desiredState = ...
-func WaitForBenthosManagerMultiState(
-	ctx context.Context,
-	manager *benthosfsm.BenthosManager,
-	fullCfg config.FullConfig,
-	desiredMap map[string]string, // e.g. { "svc1": "idle", "svc2": "active" }
-	maxAttempts int,
-	startTick uint64,
-) (uint64, error) {
-	tick := startTick
-	for i := 0; i < maxAttempts; i++ {
-		err, _ := manager.Reconcile(ctx, fullCfg, tick)
-		if err != nil {
-			return tick, err
-		}
-		tick++
-
-		allMatched := true
-		for svc, desired := range desiredMap {
-			inst, found := manager.GetInstance(svc)
-			if !found || inst.GetCurrentFSMState() != desired {
-				allMatched = false
-				break
-			}
-		}
-		if allMatched {
-			return tick, nil
-		}
-	}
-	return tick, fmt.Errorf("not all instances reached desired states after %d attempts", maxAttempts)
-}
-
 // SetupBenthosInstance creates and configures a Benthos instance for testing.
 // Returns the instance, the mock service, and the config used to create it.
 func SetupBenthosInstance(serviceName string, desiredState string) (*benthosfsm.BenthosInstance, *benthossvc.MockBenthosService, config.BenthosConfig) {
@@ -435,32 +331,6 @@ func ResetBenthosInstanceError(mockService *benthossvc.MockBenthosService) {
 	mockService.ForceRemoveBenthosError = nil
 }
 
-// SetupServiceInManager adds a service to the manager and configures it with the mock service.
-// It creates an instance automatically and adds it to the manager.
-// Note: Since this cannot directly set the service field in the instance after creation,
-// the instance in the manager will be using the default service, not the mock service.
-// Tests should rely on the manager.Reconcile method instead, which uses the manager's own service.
-func SetupServiceInManager(
-	manager *benthosfsm.BenthosManager,
-	mockService *benthossvc.MockBenthosService,
-	serviceName string,
-	desiredState string,
-) {
-	// Create a properly configured instance
-	instance := benthosfsm.NewBenthosInstance(constants.S6BaseDir, CreateBenthosTestConfig(serviceName, desiredState))
-
-	// Add it to the manager
-	manager.BaseFSMManager.AddInstanceForTest(serviceName, instance)
-
-	// Make sure the service exists in the mock service
-	mockService.ExistingServices[serviceName] = true
-
-	// Ensure we have a service info initialized
-	if mockService.ServiceStates[serviceName] == nil {
-		mockService.ServiceStates[serviceName] = &benthossvc.ServiceInfo{}
-	}
-}
-
 // CreateMockBenthosInstance creates a Benthos instance for testing.
 // Note: This creates a standard instance without replacing the service component.
 // In actual tests, the pattern is to use the manager's functionality rather than
@@ -468,34 +338,6 @@ func SetupServiceInManager(
 func CreateMockBenthosInstance(serviceName string, mockService benthossvc.IBenthosService, desiredState string) *benthosfsm.BenthosInstance {
 	cfg := CreateBenthosTestConfig(serviceName, desiredState)
 	return benthosfsm.NewBenthosInstance(constants.S6BaseDir, cfg)
-}
-
-// CreateMockBenthosManager creates a Benthos manager with a mock service for testing
-func CreateMockBenthosManager(name string) (*benthosfsm.BenthosManager, *benthossvc.MockBenthosService) {
-	// Use the specialized constructor that sets up fully mocked services
-	manager, mockService := benthosfsm.NewBenthosManagerWithMockedServices(name)
-
-	// Set up the mock service to prevent real filesystem operations
-	s6MockService := mockService.S6Service.(*s6svc.MockService)
-
-	// Configure default responses to prevent real filesystem operations
-	s6MockService.CreateError = nil
-	s6MockService.RemoveError = nil
-	s6MockService.StartError = nil
-	s6MockService.StopError = nil
-	s6MockService.ForceRemoveError = nil
-
-	// Set up the mock to say services exist after creation
-	s6MockService.ServiceExistsResult = true
-
-	// Configure default successful statuses
-	s6MockService.StatusResult = s6svc.ServiceInfo{
-		Status: s6svc.ServiceUp,
-		Pid:    12345, // Fake PID
-		Uptime: 60,    // Fake uptime in seconds
-	}
-
-	return manager, mockService
 }
 
 // StabilizeBenthosInstance ensures the Benthos instance reaches and remains in a stable state.
@@ -540,35 +382,4 @@ func StabilizeBenthosInstance(
 	return tick, fmt.Errorf(
 		"failed to reach state %s after %d attempts; current state: %s",
 		targetState, maxAttempts, instance.GetCurrentFSMState())
-}
-
-// ConfigureBenthosManagerForState sets up the mock service in a BenthosManager to facilitate
-// a state transition for a specific instance. This should be called before starting
-// reconciliation if you want to ensure state transitions happen correctly.
-//
-// Parameters:
-//   - mockService: The mock service from the manager
-//   - serviceName: The name of the service to configure
-//   - targetState: The desired state to configure the service for
-func ConfigureBenthosManagerForState(
-	mockService *benthossvc.MockBenthosService,
-	serviceName string,
-	targetState string,
-) {
-	// Make sure the service exists in the mock
-	if mockService.ExistingServices == nil {
-		mockService.ExistingServices = make(map[string]bool)
-	}
-	mockService.ExistingServices[serviceName] = true
-
-	// Make sure service state is initialized
-	if mockService.ServiceStates == nil {
-		mockService.ServiceStates = make(map[string]*benthossvc.ServiceInfo)
-	}
-	if mockService.ServiceStates[serviceName] == nil {
-		mockService.ServiceStates[serviceName] = &benthossvc.ServiceInfo{}
-	}
-
-	// Configure the service for the target state
-	TransitionToBenthosState(mockService, serviceName, targetState)
 }
